Metis CMake config: set(CMAKE_MACOSX_RPATH TRUE)
With CMake 3.10.1 out of the box I get warnings about a policy change
wrt the default value of this variable.  Setting it explicitly
squashes the warning.

IMHO there would be fewer warnings about policy changes and you'd be
able to take advantage of new policies and claimed better behavior on
the Mac if you upped the cmake_minimum_required version.  But that
choice also depends on the users you are targeting, and I cannot speak
to that.