Use @rpath and -rpath effectively on MacOS
A nice default install_name for a MacOS shared library is
@rpath/libcolamd.2.dylib instead of just libcolamd.2.dylib.  The
former method gives executables and libraries that depend on libcolamd
a good hint about how to find colamd: use the RPATH list in the
executable or dependent library.  This is the technique used by CMake
(on my system with your CMake settings) for the library in the build
tree.
Without the -rpath $(INSTALL_LIB) setting on the build of the Demo
programs, I get failures: cannot find libcolamd.2.dylib.