SuiteSparseQR_C.h: Don't define Complex
The preprocessor macro `Complex` was defined to types with possibly
different alignment and size depending on whether this header is included
in C or C++ code. This could lead to issues if it is used for function
arguments.
That macro doesn't seem to be used currently. So, just remove it.