CXSparse: Avoid including C header in C++ code
According to [1]: "Source files that are not intended to also be valid
ISO C should not use any of the C headers."

It is still correct to define `cs_complex_t` as `double _Complex` in C
and C++ code. But the C standard header `complex.h` shouldn't be included
in a header that can be used in C++ code. That is not a requirement. But
it is a recommendation.

Additionally, avoid inclusion of headers inside an `extern "C"` block.

[1]: https://en.cppreference.com/w/cpp/standard_library