SPQR: Overhaul CMake build rules.
Use targets more consistently instead of variables.
Set properties on the target instead of globally where possible.
No need for a CUDA compiler for the sources of SPQR.
There are no OpenMP accelerated expressions in the sources. So, don't depend on OpenMP.
Don't link transient targets of CHOLMOD (AMD and COLAMD) explicitly.
Link static libraries to static libraries if possible.