LAGraph: Use compile-time checks for size of uintptr_t
Instead of checking the size of `uintptr_t` on run-time (inside a loop),
move that check to the preprocessor. That way code that will never be
executed isn't compiled.

Modern compilers might be able to optimize the checks and code paths
away that will never be executed, there is no guarantees they will do
that.

I hope it is ok to check at compile time if the pointer size is neither
32-bit nor 64-bit. (Compilation will likely fail at other points anyway
if that should be the case.)