SuiteSparse_config: Check if linking to librt is necessary
On some platform, it is necessary to link to librt when using
`clock_gettime`.  Add configure checks to detect if the function is
available and whether it needs linking with librt.

That facilitates linking statically to libsuitesparseconfig using the
CMake import targets or the pkg-config files.  It also helps to avoid
issues when linking to the shared library with `--as-needed`.