CI: "Fake" a cross-compilation for one of the Ubuntu runners.
When cross-compiling, some files (notably the JIT package generator of
GraphBLAS) are built in a different order and with different commands
than normally.

Setting CMAKE_SYSTEM_NAME to the host system name doesn't actually
lead to a cross-compilation. But it should trigger the build path that
would also be used for an actual cross-compilation.