CHOLMOD: Add preprocessor flags that affect cholmod.h to targets and .pc
Some preprocessor flags that are conditionally set by the build rules of
CHOLMOD affect which symbols are exported from the built libraries.
Add the preprocessor flags that are used to guard function declarations
in `cholmod.h` to the CMake targets and the pkg-config file. That way
consumers of the library get only those declarations that are actually
exported from the library when including `cholmod.h` in their code.

Fixes #794.