CI (MSVC): Update build rules for MSVC 19.40
GitHub seems to be updating their runner images to include MSVC 19.40
currently.

NVIDIA CUDA seems to think this newer version of MSVC is no longer
Microsoft Visual Studio 2022 (but it is).
Work around that issue by adding `-allow-unsupported-compiler` to the
CUDA flags.

Additionally, it looks like the PATH variable on these runner images is
arranged such that a different (older) `clang-cl` is being picked up
by CMake instead of the newer one from MSVC 19.40.
Work around that by moving that version out of the PATH so CMake picks
the compiler that is installed with Microsoft Visual Studio.