AMD: Avoid stubs in favor of object libraries.
Avoid the stub files that were used to build objects for the int64_t
interface. Instead, use object libraries that build the same source files
once with `DLONG` and once without it.

This also avoids building all objects twice (for the shared and the
static library), effectively reducing the build time approximately by a
factor of two.