CMake: Adjust condition for changed name for static libraries.
MSVC-like compilers use the same default suffix for static libraries and
import libraries.
Instead of checking whether a compiler is MSVC or simulates it, directly
check whether these default suffices clash.

That potentially fixes issue #895. (And even in case it does not, the
updated condition is much more to the point.